feat: add 'is_allowlisted' attribute to job serializer
[APER-2377]

I am requesting an update to the data provided by the JobSerializer. We would like to limit the job options a learner can select from using the B2C version of the Skills Builder. To do this, we need an attribute in the data pushed to Algolia that we can filter on.

To do this, we have added a new derived field to the JobSerializer. It will check a new setting (`TAXONOMY_JOB_ID_ALLOWLIST`) which is a (string) list of job ids (the lightcast `external_id`). If the external ID of a job matches an item on the list, it is "allowlisted" and can be displayed as an option to the B2C learners.

I don't think this change should affect existing consumers of this Algolia index.

Thanks for your consideration!